Correction and notes
Correction: It makes sense to set the start location at the current location (btw, this is the "local" location) if the simulator preserves this information from the previous flight. Which it does, judging by the behavior from multiple consecutive runs of the zig-zag trajectory with the starter code.

Notes: Implementation notes. Labeled TODOs for v.2 (post-submission).